+++ tests/data_import/__init__.py
@@ -1,0 +1,3 @@
+# NOTE: test_insert_sessions does not increase coverage over common/test_insert
+# but it does declare it's own nwbfile without downloading and test broken
+# links which aren't technically part of spyglass